Move Vfs and SharedPaths into util/doc_helpers
